Refactor some type utilities
The ability to use an xs:anyURI as the value of an xs:string option necessitated
moving some things around. I still think the whole type checking area needs a rewrite.
